+++ packages/fluent/src/TypeOrmConnector/test/sqlite/TypeOrmSqliteConnector.spec.ts
@@ -17,6 +17,15 @@ beforeAll(async () => {
   await Fluent.initialize([MemoryDataSource], dbEntities)
 })
 
+const userRepo = new UserRepository()
+describe('Loading test', () => {
+  it('Should run even when initialized in the same file', async () => {
+    const a = await userRepo.findMany()
+
+    expect(Array.isArray(a)).toBe(true)
+  })
+})
+
 describe('Execute all basic test Suite', () => {
   basicTestSuite(GoatRepository)
 })